+++ app/src/main/java/com/depromeet/threedays/front/domain/usecase/member/GetMemberUseCase.java
@@ -3,6 +3,7 @@ package com.depromeet.threedays.front.domain.usecase.member;
 import com.depromeet.threedays.data.entity.member.MemberEntity;
 import com.depromeet.threedays.data.enums.MemberStatus;
 import com.depromeet.threedays.front.config.security.AuditorHolder;
+import com.depromeet.threedays.front.config.security.filter.token.TokenGenerator;
 import com.depromeet.threedays.front.domain.converter.member.MemberConverter;
 import com.depromeet.threedays.front.domain.model.member.Member;
 import com.depromeet.threedays.front.domain.model.member.SaveMemberUseCaseResponse;
@@ -10,7 +11,6 @@ import com.depromeet.threedays.front.domain.model.member.Token;
 import com.depromeet.threedays.front.domain.query.GetMemberQuery;
 import com.depromeet.threedays.front.exception.MemberNotFoundException;
 import com.depromeet.threedays.front.persistence.repository.member.MemberRepository;
-import com.depromeet.threedays.front.support.TokenGenerator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
