+++ app/src/main/java/com/depromeet/threedays/front/domain/usecase/member/GetTokenUseCase.java
@@ -1,10 +1,10 @@
 package com.depromeet.threedays.front.domain.usecase.member;
 
+import com.depromeet.threedays.front.config.security.filter.token.TokenGenerator;
 import com.depromeet.threedays.front.config.security.filter.token.TokenResolver;
 import com.depromeet.threedays.front.domain.model.member.Token;
 import com.depromeet.threedays.front.exception.RefreshTokenInvalidException;
 import com.depromeet.threedays.front.persistence.repository.member.MemberRepository;
-import com.depromeet.threedays.front.support.TokenGenerator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
