+++ app/src/main/java/com/depromeet/threedays/front/domain/usecase/member/SaveMemberUseCase.java
@@ -1,12 +1,12 @@
 package com.depromeet.threedays.front.domain.usecase.member;
 
 import com.depromeet.threedays.data.entity.member.MemberEntity;
+import com.depromeet.threedays.front.config.security.filter.token.TokenGenerator;
 import com.depromeet.threedays.front.domain.command.SaveMemberCommand;
 import com.depromeet.threedays.front.domain.converter.member.MemberConverter;
 import com.depromeet.threedays.front.domain.model.member.SaveMemberUseCaseResponse;
 import com.depromeet.threedays.front.domain.model.member.Token;
 import com.depromeet.threedays.front.persistence.repository.member.MemberRepository;
-import com.depromeet.threedays.front.support.TokenGenerator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
